refactor(drawer): move WithGuideCue story with play test
